Move baseline utils to the baseline package (#2738)
* Move detekt result wrapper to baseline package

* Move utility to baseline package where it is only used

* Remove obsolete MessageCollector.kt
