--- detekt-cli/src/main/kotlin/io/gitlab/arturbosch/detekt/cli/MessageCollector.kt
@@ -1,5 +1,0 @@
-package io.gitlab.arturbosch.detekt.cli
-
-interface MessageCollector {
-    operator fun plusAssign(msg: String)
-}
