+++ detekt-cli/src/main/kotlin/io/gitlab/arturbosch/detekt/cli/baseline/BaselineFilteredResult.kt
@@ -1,16 +1,18 @@
-package io.gitlab.arturbosch.detekt.cli
+package io.gitlab.arturbosch.detekt.cli.baseline
 
 import io.gitlab.arturbosch.detekt.api.Detektion
 import io.gitlab.arturbosch.detekt.api.Finding
 import io.gitlab.arturbosch.detekt.api.RuleSetId
-import io.gitlab.arturbosch.detekt.cli.baseline.BaselineFacade
 
-class FilteredDetectionResult(detektion: Detektion, baselineFacade: BaselineFacade) : Detektion by detektion {
+class BaselineFilteredResult(
+    result: Detektion,
+    baselineFacade: BaselineFacade
+) : Detektion by result {
 
     private val filteredFindings: Map<RuleSetId, List<Finding>>
 
     init {
-        filteredFindings = detektion.findings
+        filteredFindings = result.findings
                 .map { (key, value) -> key to baselineFacade.filter(value) }
                 .toMap()
     }
